+++ elide-graphql/src/test/groovy/com/yahoo/elide/graphql/GraphQLEndointTest.groovy
@@ -21,6 +21,7 @@ import graphqlEndpointTestModels.DisallowShare
 import graphqlEndpointTestModels.security.CommitChecks
 import graphqlEndpointTestModels.security.UserChecks
 import org.mockito.Mockito
+import org.skyscreamer.jsonassert.JSONAssert
 import org.testng.Assert
 import org.testng.annotations.BeforeMethod
 import org.testng.annotations.BeforeTest
@@ -172,10 +173,66 @@ class GraphQLEndointTest {
                         }
                       ]
                     }
-                  }
+                  },
+                  "extensions": null
                 }
                 '''
         def response = endpoint.post(user1, graphQLRequestToJSON(graphQLRequest))
+        assert200EqualBody(response, graphQLResponse)
+    }
+
+    @Test
+    void testValidFetchWithVariables() {
+        String graphQLRequest =
+            '''
+            query myQuery($bookId: [String]) {
+              book(ids: $bookId) {
+                edges {
+                  node {
+                    id
+                    title
+                    authors {
+                      edges {
+                        node {
+                          name
+                        }
+                      }
+                    }
+                  }
+                }
+              }
+            }
+            '''
+        String graphQLResponse =
+                '''
+                {
+                  "errors": [],
+                  "data": {
+                    "book": {
+                      "edges": [
+                        {
+                          "node": {
+                            "id": "1",
+                            "title": "My first book",
+                            "authors": {
+                              "edges": [
+                                {
+                                  "node": {
+                                    "name": "Ricky Carmichael"
+                                  }
+                                }
+                              ]
+                            }
+                          }
+                        }
+                      ]
+                    }
+                  },
+                  "extensions": null
+                }
+                '''
+        def variables = ["bookId": "1"]
+        def response = endpoint.post(user1, graphQLRequestToJSON(graphQLRequest, variables))
         assert200EqualBody(response, graphQLResponse)
     }
 
@@ -207,7 +264,8 @@ class GraphQLEndointTest {
                         }
                       ]
                     }
-                  }
+                  },
+                  "extensions": null
                 }
                 '''
         def response = endpoint.post(user1, graphQLRequestToJSON(graphQLRequest))
@@ -322,7 +380,8 @@ class GraphQLEndointTest {
                         }
                       ]
                     }
-                  }
+                  },
+                  "extensions": null
                 }
                 '''
         response = endpoint.post(user2, graphQLRequestToJSON(graphQLRequest))
@@ -416,7 +475,8 @@ class GraphQLEndointTest {
                         }
                       ]
                     }
-                  }
+                  },
+                  "extensions": null
                 }
                 '''
 
@@ -458,7 +518,8 @@ class GraphQLEndointTest {
                         }
                       ]
                     }
-                  }
+                  },
+                  "extensions": null
                 }
                 '''
 
@@ -526,7 +587,8 @@ class GraphQLEndointTest {
                         }
                       ]
                     }
-                  }
+                  },
+                  "extensions": null
                 }
                 '''
 
@@ -595,7 +657,8 @@ class GraphQLEndointTest {
                       }
                     ]
                   }
-                }
+                },
+                "extensions": null
               }
               '''
         response = endpoint.post(user1, graphQLRequestToJSON(graphQLRequest))
@@ -623,29 +686,35 @@ class GraphQLEndointTest {
     }
 
     static String graphQLRequestToJSON(String request) {
+        return graphQLRequestToJSON(request, new HashMap<String, String>())
+    }
+
+    static String graphQLRequestToJSON(String request, Map<String, String> variables) {
+        ObjectMapper mapper = new ObjectMapper();
         JsonNode node = JsonNodeFactory.instance.objectNode()
         node.put("query", request)
+        node.set("variables", variables == null ? null : mapper.valueToTree(variables))
         return node.toString()
     }
 
     static JsonNode extract200Response(Response response) {
+        return new ObjectMapper().readTree(extract200ResponseString(response))
+    }
+
+    static String extract200ResponseString(Response response) {
         Assert.assertEquals(response.getStatus(), 200)
-        ObjectMapper mapper = new ObjectMapper()
-        return mapper.readTree((String) response.getEntity())
+        return (String) response.getEntity()
     }
 
     static assert200EqualBody(Response response, String expected) {
-        ObjectMapper mapper = new ObjectMapper()
-        JsonNode expectedNode = mapper.readTree(expected)
-        JsonNode actualNode = extract200Response(response)
-        Assert.assertEquals(actualNode, expectedNode)
+        String actual = extract200ResponseString(response)
+        JSONAssert.assertEquals(expected, actual, true)
     }
 
     static assert200DataEqual(Response response, String expected) {
-        ObjectMapper mapper = new ObjectMapper()
-        JsonNode expectedNode = mapper.readTree(expected)
         JsonNode actualNode = extract200Response(response)
-        Assert.assertEquals(actualNode.get("data"), expectedNode)
+        String actual = new ObjectMapper().writeValueAsString(actualNode.get("data"))
+        JSONAssert.assertEquals(expected, actual, true)
     }
 
     static assertHasErrors(Response response) {