+++ changelog.md
@@ -6,10 +6,12 @@
  * Add additional logging around exception handling.
  * Don't swallow generic Exception in Elide. Log it and bubble it up to caller.
  * Fix a bug where null filter expressions were possible if no filter was passed in by the user, but permission filters existed.
+ * Fix support for handling GraphQL variables.
 
 **Features**
  * Handle ConstraintViolationException's by extracting the first constraint validation failure.
  * Include GraphQL in Elide standalone by default with ability to remove it via dependency management.
+ * Upgrade to the latest graphql-java version: 6.0.
 
 ## 4.0-beta-3
 **Fixes**